chore: remove unused rewards methods
